+++ src/js/engagement_view/src/services/fetch.ts
@@ -1,4 +1,4 @@
-const apiFetch = async (urlSlug: string, method = "get") => {
+export const apiFetch = async (urlSlug: string, method = "get") => {
   // const url = `api/${encodeURIComponent(urlSlug)}`;
 
   const response = await fetch(urlSlug, {
@@ -15,4 +15,22 @@ const apiFetch = async (urlSlug: string, method = "get") => {
   return response.json();
 };
 
-export default apiFetch;
+
+// export default apiFetch;
+
+
+export const apiPostFetch = async(urlSlug: string, method = "post", body:string) => {
+  const response = await fetch(urlSlug, {
+    method,
+    credentials: "include",
+    headers: new Headers({
+      "Content-Type": "application/json",
+    }),
+    body: body,
+  }).catch((e) => {
+    console.warn(e);
+    throw new Error(`Error: ${e}`);
+  });
+
+  return response.json();
+};
