+++ src/js/engagement_view/src/services/graphQL/utils_GraphQlEndpointRetrieveGraph/expandScopeQuery.tsx
