+++ src/js/engagement_view/src/services/login/checkLoginService.ts
@@ -1,5 +1,5 @@
-import apiFetch from "./fetch";
-import DEV_API_EDGES from "./constants";
+import { apiFetch } from "../fetch";
+import DEV_API_EDGES from "../constants";
 
 export const checkLogin = async (): Promise<boolean | null> => {
   try {
