+++ src/js/engagement_view/src/services/sagemakerNotebook/notebookService.ts
@@ -1,5 +1,5 @@
-import apiFetch from "./fetch";
-import DEV_API_EDGES from "./constants";
+import {apiFetch} from "../fetch";
+import DEV_API_EDGES from "../constants";
 
 export const getNotebookUrl = async (): Promise<string> => {
   return apiFetch(`${DEV_API_EDGES.auth}/getNotebook`, "post").then(
